envoy/eds: move EDS related functionality into its pkg
Cluster load assignment is a part of EDS, so there's no
reason for this to be its own package. This was just
left behind as is from the initial code base that
had all the XDS components split as separate services.
