--- pkg/envoy/cla/suite_test.go
@@ -1,13 +1,0 @@
-package cla
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestClusterLoadAssignment(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Cluster Load Assignment Test Suite")
-}
