--- pkg/envoy/cla/types.go
@@ -1,10 +1,0 @@
-// Package cla implements Envoy's cluster load assignment functionality for discovered endpoints.
-package cla
-
-import (
-	"github.com/openservicemesh/osm/pkg/logger"
-)
-
-var (
-	log = logger.New("envoy/cla")
-)
